Create Climbing stairs(70)
The task is to find all the ways to get to the top. Steps can be with the length of 1 or 2. 
P.S. :
I tried to find the ways by recurtion but it didn't seem to be fast enough. I also noticed that the result seems to be same as Fibonacci number, so i used it to find the answer